Adds app icon and fix architecture
