+++ CavernSamples/CavernizeGUI/App.axaml.cs
@@ -1,6 +1,7 @@
 ﻿using Avalonia;
 using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Markup.Xaml;
+using Avalonia.Threading;
 
 namespace CavernizeGUI;
 
@@ -9,14 +10,18 @@ public partial class App : Application {
 
     public override void OnFrameworkInitializationCompleted() {
         if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop) {
-            MainViewModel viewModel = new();
-            if (Program.Args.Length != 0) {
-                viewModel.InitializeCommandLine(Program.Args);
+            MainWindow window = new();
+            if (Program.Args.Length == 0) {
+                desktop.MainWindow = window;
+            } else {
+                bool initialized = window.InitializeCommandLine(Program.Args);
+                if (Program.ConsoleMode || !initialized) {
+                    Program.ExitCode = initialized || Program.IsHelpCommand(Program.Args) ? 0 : 1;
+                    Dispatcher.UIThread.Post(() => desktop.Shutdown(Program.ExitCode));
+                    return;
+                }
+                desktop.MainWindow = window;
             }
-
-            desktop.MainWindow = new MainWindow {
-                DataContext = viewModel
-            };
         }
 
         base.OnFrameworkInitializationCompleted();
