+++ CavernSamples/CavernizeGUI/QueuedRenderJob.cs
@@ -1,9 +1,13 @@
 ﻿using Cavernize.Logic.Models;
 using Cavernize.Logic.Models.RenderTargets;
+using System.ComponentModel;
+using System.Runtime.CompilerServices;
 
 namespace CavernizeGUI;
 
-public sealed class QueuedRenderJob : ObservableObject {
+public sealed class QueuedRenderJob : INotifyPropertyChanged {
+    public event PropertyChangedEventHandler PropertyChanged;
+
     public string SourcePath { get; }
 
     public string OutputPath {
@@ -84,4 +88,16 @@ public sealed class QueuedRenderJob : ObservableObject {
         UpmixingSmoothness = upmixingSmoothness;
     }
 
+    void OnPropertyChanged([CallerMemberName] string propertyName = null) =>
+        PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+
+    bool SetProperty<T>(ref T field, T value, [CallerMemberName] string propertyName = null) {
+        if (EqualityComparer<T>.Default.Equals(field, value)) {
+            return false;
+        }
+
+        field = value;
+        OnPropertyChanged(propertyName);
+        return true;
+    }
 }
